feat: refactore library utilities
